Se modifcó interfaz gráfica agregando excepciones
